+++ sdk/spring/azure-spring-boot/CHANGELOG.md
@@ -7,7 +7,7 @@
 ### Breaking Changes
 
 ### Bugs Fixed
-- Fix bug that application type validation. ([#23481](https://github.com/Azure/azure-sdk-for-java/issues/23481))
+- Fix bug about application type validation. ([#23481](https://github.com/Azure/azure-sdk-for-java/issues/23481))
 
     The latest application type relationship is as follows:
     
@@ -17,8 +17,10 @@
     |                      No                       |                          Yes                           |                       `resource_server`                                                               |       `resource_server`     |
     |                      Yes                      |                          Yes                           | `web_application`,`resource_server`,`resource_server_with_obo`, `web_application_and_resource_server` | `resource_server_with_obo`  |
 
+### Deprecations
+- Deprecate `azure.activedirectory.authorization-clients.xxx.on-demand`, use `azure.activedirectory.authorization-clients.xxx.authorization-grant-type` instead. If you are using `on-demand=true`, please change to `authorization-grant-type=authorization_code`.
+
 ### Other Changes
-- Deprecate `com.azure.spring.aad.webapp.AuthorizationClientProperties.onDemand`, use `com.azure.spring.aad.webapp.AuthorizationClientProperties.authorizationGrantType` is enough.
 
 
 ## 3.8.0 (2021-08-25)